-Ticket #2744
git-svn-id: https://subversion.assembla.com/svn/mdss@7009 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 6adccccbf30ad07f9bb3b32f4ef9f147db971dad [formerly 1d9abdd454d280734036199ef8b7239d47744685]
Former-commit-id: d8cddadd349e1ce87955910fd1f25e1428909084

+++ trunk/src/server/dss/vector/solutions/generator/MultiTermListener.java
@@ -98,7 +98,7 @@ public class MultiTermListener extends ExcelAdapter implements ExcelExportListen
           }
           else
           {
-            if (relationship != null)
+            if (relationship != null && !relationship.isNew())
             {
               relationship.delete();
             }
